feat: visibility and javascript improvement
