+++ docs/script.js
@@ -1,3 +1,11 @@
+function unEscape(htmlStr) {
+    htmlStr = htmlStr.replace(/&lt;/g , "<");	 
+    htmlStr = htmlStr.replace(/&gt;/g , ">");     
+    htmlStr = htmlStr.replace(/&quot;/g , "\"");  
+    htmlStr = htmlStr.replace(/&#39;/g , "\'");   
+    htmlStr = htmlStr.replace(/&amp;/g , "&");
+    return htmlStr;
+}
 //(function (document) {
 function onceLoaded(document) {
 	//http://stackoverflow.com/a/10372280/398634
@@ -155,6 +163,8 @@ function onceLoaded(document) {
 			fit: true,
 			center: true,
 		});
+		// include script from the SVG
+		eval(unEscape(document.getElementById('SVGScript').innerHTML));
 
 		//updateState()
 	}
