+++ examples/webassembly/server/assets/index.html
@@ -6,17 +6,6 @@
 		<link rel="stylesheet" type="text/css" href="style.css" media="screen" />
 		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/github-fork-ribbon-css/0.2.3/gh-fork-ribbon.min.css" />
 
-		<script src="wasm_exec.js"></script>
-		<script>
-			const go = new Go();
-			WebAssembly.instantiateStreaming(fetch("wtg.wasm"), go.importObject)
-				.then((result) => {
-					go.run(result.instance);
-				})
-				.then(instance => {
-					onceLoaded(document);
-				});
-		</script>
 	</head>
 	<body>
 		<pre id="editor">
@@ -114,6 +103,17 @@ stage4: commodity / accepted
 
 		<script src="script.js"></script>
 		<script src="svg-pan-zoom.min.js" type="text/javascript" charset="utf-8"></script>
+		<script src="wasm_exec.js"></script>
+		<script>
+			const go = new Go();
+			WebAssembly.instantiateStreaming(fetch("wtg.wasm"), go.importObject)
+				.then((result) => {
+					go.run(result.instance);
+				})
+				.then(instance => {
+					onceLoaded(document);
+				});
+		</script>
 
 	</body>
 </html>
